renamed application.py to service.py
